Garapena/Pilotoak/EkozirAPP/requirements.txt
Flask==3.0.3
web3==6.15.1
python-dotenv==1.0.1
eth-typing==5.0.0
gunicorn==21.2.0

